+++ wulttools/ndl/_NdlReport.py
@@ -48,5 +48,5 @@ def report_command(args):
     rep = NdlReport.NdlReport(rsts, args.outdir, report_descr=args.report_descr,
                               xaxes=args.xaxes, yaxes=args.yaxes, hist=args.hist,
                               chist=args.chist, logpath=logpath)
-    rep.relocatable = args.relocatable
+    rep.copy_raw= args.copy_raw
     rep.generate()