+++ wulttools/wult/_Wult.py
@@ -180,7 +180,7 @@ def _build_arguments_parser():
 
     subpars.add_argument("--reportids", help=_Common.REPORTIDS_DESCR)
     subpars.add_argument("--report-descr", help=_Common.REPORT_DESCR)
-    subpars.add_argument("--relocatable", action="store_true", help=_Common.RELOCATABLE_DESCR)
+    subpars.add_argument("--copy-raw", action="store_true", help=_Common.COPY_RAW_DESCR)
     subpars.add_argument("--list-metrics", action="store_true", help=_Common.LIST_METRICS_DESCR)
 
     text = """Generate HTML report with a pre-defined set of diagrams and histograms. Possible