+++ wulttools/wult/_WultStart.py
@@ -67,7 +67,7 @@ def _generate_report(args):
 
     rsts = _Common.open_raw_results([args.outdir], args.toolname)
     rep = WultReport.WultReport(rsts, args.outdir / "html-report", report_descr=args.reportid)
-    rep.relocatable = False
+    rep.copy_raw = False
     rep.set_hover_metrics(_WultCommon.HOVER_METRIC_REGEXS)
     rep.generate()